face-recognition/face-api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
torch==2.2.2
torchvision==0.17.2
transformers==4.43.2
timm==1.0.19
pillow==10.2.0
numpy==1.26.4
pydantic==2.11.7
python-multipart==0.0.20
huggingface-hub==0.34.4
